refactor: rename seasons-service to seasons-utils
refactor: move relevant functions from utils to season-utils
